Text Localization
Added the @nuxtjs/i18n package and configured It in the i18n.config.ts file, as well in the nuxt.config.ts.

Created a locales folder that holds yaml files with the translation keys.

Added the locales location to the .vscode settings file, under the i18n-ally extension.

Added tailwindcss and i18n-ally to the recommended extensions.